# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SokoOS Cloud API
  version: 1.0.0
  description: |
    NestJS modular monolith REST API for SokoOS.
    All business routes are under `/v1`. Health probes are unversioned.
servers:
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /v1/returns:
    get:
      tags: [Sales]
      summary: List returns
      parameters:
        - in: query
          name: saleId
          schema: { type: string, format: uuid }
        - in: query
          name: branchId
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Returns page
    post:
      tags: [Sales]
      summary: Create return (partial lines + compensating stock movements)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [saleId, items]
              properties:
                saleId: { type: string, format: uuid }
                warehouseId: { type: string, format: uuid }
                reason: { type: string }
                items:
                  type: array
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId: { type: string, format: uuid }
                      quantity: { type: integer, minimum: 1 }
                      unitPriceMinor: { type: integer }
      responses:
        "200":
          description: Completed return
  /v1/returns/{id}:
    get:
      tags: [Sales]
      summary: Get return by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Return detail
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []